Regrouped accessor internals interface and class in same file
